# nest/src/tasks/tasks.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  Param,
  ParseIntPipe,
  ParseUUIDPipe,
  Patch,
  Post,
  Query,
  UseGuards,
} from '@nestjs/common';
import { CurrentUser } from '../auth/current-user.decorator';
import { JwtAuthGuard } from '../auth/jwt-auth.guard';
import type { AuthenticatedUser } from '../auth/jwt.strategy';
import { ProjectAccessGuard } from '../auth/project-access.guard';
import { BulkUpdateTasksDto } from './dto/bulk-update-tasks.dto';
import { CreateTaskDto } from './dto/create-task.dto';
import { TaskFilterDto } from './dto/task-filter.dto';
import { UpdateTaskDto } from './dto/update-task.dto';
import { TasksService } from './tasks.service';

@Controller('projects/:projectId/tasks')
@UseGuards(JwtAuthGuard, ProjectAccessGuard)
export class TasksController {
  constructor(private readonly tasks: TasksService) {}

  @Get()
  list(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Query() filter: TaskFilterDto,
  ) {
    return this.tasks.listByProject(user.tenantId, projectId, filter);
  }

  @Get('count')
  async count(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Query() filter: TaskFilterDto,
  ) {
    return { count: await this.tasks.count(user.tenantId, projectId, filter) };
  }

  // ':id' より前に宣言する（後だと 'by-seq' が :id の ParseUUIDPipe に捕まり 400 になる）。
  // 一覧に載らないタスク（完了済み等）を ?task=seq のディープリンク/通知から開くための単体解決。
  @Get('by-seq/:seq')
  findBySeq(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Param('seq', ParseIntPipe) seq: number,
  ) {
    return this.tasks.findBySeqInProject(user.tenantId, projectId, seq);
  }

  @Get(':id')
  find(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Param('id', new ParseUUIDPipe()) id: string,
  ) {
    return this.tasks.findInProject(user.tenantId, projectId, id);
  }

  @Get(':id/activities')
  activities(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Param('id', new ParseUUIDPipe()) id: string,
  ) {
    return this.tasks.listActivities(user.tenantId, projectId, id);
  }

  @Post()
  create(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Body() dto: CreateTaskDto,
  ) {
    return this.tasks.create(user.tenantId, projectId, dto, user.userId);
  }

  // ':id' より前に宣言する（後だと 'bulk' が :id の ParseUUIDPipe に捕まり 400 になる）
  @Patch('bulk')
  bulkUpdate(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Body() dto: BulkUpdateTasksDto,
  ) {
    return this.tasks.bulkUpdate(user.tenantId, projectId, dto, user.userId);
  }

  @Patch(':id')
  update(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Param('id', new ParseUUIDPipe()) id: string,
    @Body() dto: UpdateTaskDto,
  ) {
    return this.tasks.update(user.tenantId, projectId, id, dto, user.userId);
  }

  @Delete(':id')
  @HttpCode(204)
  remove(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Param('id', new ParseUUIDPipe()) id: string,
  ) {
    return this.tasks.remove(user.tenantId, projectId, id, user.userId);
  }
}